wordnik-tool/requirements.txt
gradio
requests
python-dotenv
smolagents